finalizacao do Flight e validacoes

+++ projOnTheFly.Flights/Controllers/FlightsController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
+using projOnTheFly.Flights.DTO;
 using projOnTheFly.Flights.Service;
 using projOnTheFly.Models;
 using projOnTheFly.Passenger.DTO;
@@ -32,12 +33,13 @@ namespace projOnTheFly.Flights.Controllers
             => await _flightService.CheckFlight(flightCheck.Iata, flightCheck.Rab, flightCheck.Schedule);
 
         [HttpPost]
-        public async Task<ActionResult<FlightDTO>> Create(FlightDTO flightDTO)
+        public async Task<ActionResult<Models.FlightDTO>> Create(Models.FlightDTO flightDTO)
         {
             
             AirportDTO airport = await GetAirport.GetAirportAsync(flightDTO.Iata);
             if (flightDTO.Status == false) return BadRequest("status de vôo cancelado");            
             Aircraft aircraft = await GetAircraft.GetAircraftAsync(flightDTO.Rab);
+            if (aircraft.Company.Status == false) return BadRequest("Companhia com restrição");
 
             Flight f = new()
             {
@@ -54,18 +56,34 @@ namespace projOnTheFly.Flights.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult<Flight>> Update(Flight flight)
+        public async Task<ActionResult> Update(string iata, string rab, DateTime schedule)
+        {
+
+            _flightService.Update(iata, rab, schedule);            
+
+            return StatusCode(204);
+        }
+
+
+
+
+
+
+
+        /*public async Task<ActionResult<Flight>> Update(Flight flight)
         {
             _flightService.Update( flight);
 
             return StatusCode(200);
-        }
+        }*/
 
         [HttpDelete]
         public async Task<ActionResult> Delete(string iata, string rab, DateTime schedule)
         {
-
-            _flightService.Delete(iata, rab, schedule);
+            var found = await _flightService.GetByFilters(iata, rab, schedule);
+            if (found == null) return NotFound();
+            _flightService.Delete(iata, rab, schedule);          
+           
             return StatusCode(200);
 
         }
